Backup
My local repository has a damaged file in the .git folder.
Experience thought me the best way to go is to delete the repository, clone it again, and get data from a backup to be 100% up-to-date. Especially since this also involves files in the .gitignore file.
I do not need this .rar file in my remote repository, hence its addition to .gitingore.